Add Collision Detection example and separate examples (into new Force) category (better performance)
