+++ packages/layerchart/src/routes/docs/examples/Beeswarm/+page.ts
