+++ packages/layerchart/src/routes/docs/examples/ForceGroup/dots.json
