+++ packages/layerchart/src/routes/docs/examples/ForceTree/+page.ts
@@ -1,0 +1,9 @@
+import pageSource from './+page.svelte?raw';
+
+export async function load() {
+  return {
+    meta: {
+      pageSource,
+    },
+  };
+}
